pyproject.toml
[project]
name = "crypto-scanner"
version = "0.1.0"
description = "Multi-strategy crypto technical indicator scanner with Telegram notifications"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9,<4",
    "pandas>=2.1,<3",
    "ta>=0.11,<1",
    "pydantic>=2.5,<3",
    "pydantic-settings>=2.1,<3",
    "pyyaml>=6.0,<7",
    "python-dotenv>=1.0,<2",
    "apscheduler>=3.10,<4",
    "python-telegram-bot>=21,<22",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9",
    "pytest-asyncio>=0.23,<1",
    "aioresponses>=0.7,<1",
]

[project.scripts]
crypto-scanner = "src.main:main"
